Added zooming and switching to a julia set defined at a point in the mandelbrot set.

+++ fractal_functions.c
@@ -17,8 +17,8 @@ fractalOut mandelbrot(double complex c) {
             break;
         }
     }
-    float logEnd = log(cabs(z));
-    float smoothIters = n + 1 - (log(logEnd / log(2)) / log(2));
+    double logEnd = log(cabs(z));
+    double smoothIters = n + 1 - (log(logEnd / log(2)) / log(2));
     fractalOut out = {
         .iters = n, .smoothIters = smoothIters,
         .start = c,
@@ -45,7 +45,7 @@ fractalOut julia(int P, double complex z, double complex c) {
             break;
         }
     }
-    float logEnd = log(cabs(z));
+    double logEnd = log(cabs(z));
     out.smoothIters = n + 1 - (log(logEnd / log(2)) / log(P));
     out.iters = n;
     out.end = z;
